Replace loss_fn
 - Change keras provided categorical_crossentropy by custom
   categorical crossentropy. With Keras' implementation, the training
   loss stays stucked at 6.7 mark. When we replace it with the custom
   implementation the training loss finally starts to go down.
 - Even after the training loss has gotten over the initial stuck
   zone, I see that the custom implementation still reduces training
   loss much faster.

+++ squeezenet/input_imagenet.py
@@ -140,7 +140,7 @@ class InputImagenetTrain(InputImagenetBase):
 
         # Images returned by the generator can have uneven resolution
         dataset = tf.data.Dataset.from_generator(self._generator,
-                                                 output_types=(tf.dtypes.string, tf.dtypes.int64),
+                                                 output_types=(tf.dtypes.string, tf.dtypes.float32),
                                                  output_shapes=(tf.TensorShape([]),
                                                                 tf.TensorShape([1000])))
         # Prefetch samples using a separate thread for faster file-read
@@ -231,6 +231,6 @@ class InputImagenetTrain(InputImagenetBase):
         :return: A tensor of shape (None, 1000)
         """
         # Convert labels to one-hot
-        y = tf.one_hot(indices=y - 1, depth=1000, dtype=tf.dtypes.int64)
+        y = tf.one_hot(indices=y - 1, depth=1000, dtype=tf.dtypes.float32)
 
         return y
